refactor: move `env.mjs` to src/ dir
