+++ src/env.mjs
@@ -1,4 +1,5 @@
 // @ts-check
+// Using .mjs because this will be used in next.config.mjs
 
 import { vercel } from '@t3-oss/env-core/presets';
 import { createEnv } from '@t3-oss/env-nextjs';
